+++ verify/View/VerifyListener.swift
@@ -1,6 +1,6 @@
 //
 //  VerifyListener.swift
-//
+//  verify
 //
 //  Created by Jamy Bailly on 30/07/2023.
 //
@@ -8,7 +8,7 @@
 import Combine
 
 @available(iOS 15.0, *)
-class VerifyListener : ObservableObject {
+internal class VerifyListener {
     var onReady: (() -> Void)? = nil
 	var onFinished: (() -> Void)? = nil
 	var onMessage: ((String) -> Void)? = nil
