+++ verify/WebView/VerifyWebViewDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  VerifyWebView.swift
+//  verify
+//
+//  Created by Jamy Bailly on 31/10/2023.
+//
+
+import Foundation
+
+internal protocol VerifyDelegate {
+    func onReady()
+    func onFinished()
+    func onMessage(_ message: String)
+}
